Renames tasks_routes. Wrote goal to_json.  Still passing all tests
