+++ app/__init__.py
@@ -30,7 +30,7 @@ def create_app(test_config=None):
     migrate.init_app(app, db)
 
     # Register Blueprints here
-    from .routes import tasks_bp
+    from .tasks_routes import tasks_bp
     app.register_blueprint(tasks_bp)
     from .goal_routes import goals_bp
     app.register_blueprint(goals_bp)
