+++ app/models/goal.py
@@ -5,3 +5,9 @@ class Goal(db.Model):
     id = db.Column(db.Integer, primary_key=True, autoincrement=True)
     title = db.Column(db.String)
     tasks = db.relationship("Task", back_populates="goal")
+
+
+def to_json(self):
+    return {
+        "id": self.id,
+        "title": self.title}
